[v8-platform] Store the platform in a unique_ptr. (Try 2)
We want to change the signature of {CreateDefaultPlatform} in the V8
API to return a unique_ptr instead of a raw pointer to indicate that the
caller owns the platform. With this change we prepare pdfium for this
change.

In this second attempt, keep the old InitializeV8ForPDFium() method
around to complete the Chromium DEPS roll. Once that lands safely,
remove it from PDFium.

This relands commit 608e8dd6 and commit 3355f459, which commit b3a3eaab
reverts.

Change-Id: Icc60b17ca202637d34ae242c0785d939194d0fe6