Update patch set 8
Patch Set 8:

If I understand things correctly (which is a big if), it's typical for coordinates to be expressed in a world where there are 72 units per inch. That means that a single place to the right of the decimal point is enough to capture differences of 1/720th of an inch, which is beyond the resolution of a laser printer. So, how about indeed moving this code into pdfium, and modifying it to emit %.2f (giving an order of magnitude margin of precision)?

Patch-set: 8