Change return value of GetAvailInput
This changes the return value from uint32_t to FX_FILESIZE, which is
the type the methods is uses return. The existing code does na
unguarded static cast, so something like -1 could cause a very large
value being returned.

This change has a cascading impact up to the top of the progressive
codec, which now has to handle negative values gracefully.

Change-Id: I813fb71e932dd5da014dbaed0dbf3bb28f8d4e9f