Fix lifetime probe issue in CJX_Object
This CL removes the UnownedPtr to the CXFA_LayoutItem from CJX_Object.
This is because the CJX_Object will be destroyed by the CXFA_Node which
is destroyed in the CXFA_Document destructor (due to the vector of
unique_ptr being destroyed). The CXFA_LayoutItem will be freed in the
LayoutProcessor which also lives in the CXFA_Document.

Bug: chromium:807215
Change-Id: I86040e154ee2e5d461fc4d3565a10a9181680207