Add limit to number of formcalc expressions
Currently it's possible to create a formcalc script which creates a
large number of expressions. This will eventually cause stack exhaustion
as we try to allocate the needed expression objects.

This CL limits the number of parsed expressions in the PostExpression
section in order to keep from failing due to stack overflow.

Bug: chromium:799721
Change-Id: I69fca35db7f75ef97aec21c22fc06d926dfe2df6