Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> This looks reasonable to me. Leaving for Tom to approve in case he has objections/alternates

Can we get an overarching proposal for how we are going to use namespaces in PDFium before we jump off the end of this pier?  I've wanted this for a while, but we should kick it around some before impl.

Patch-set: 1
Reviewer: Gerrit User 5495 <5495@fca2d64a-0420-3532-b200-12848c95a47b>