Revert "Actually enable -Wimplicit-fallthrough for pdfium."
This reverts commit 60027a205953e95b230f289118448b55f12f9a94.

Reason for revert: This will break the fuzzer build in Chromium because the XFA side is still broken. I think we need to roll the build/ directory or clang to get the new flag and then build locally before re-enabling?

Or, build with pdf_enable_xfa = true in a chrome checkout.


Original change's description:
> Actually enable -Wimplicit-fallthrough for pdfium.
> 
> I forgot to do this in https://pdfium-review.googlesource.com/c/pdfium/+/26210
> which made PDFium compile with the warning enabled.
> 
> Bug: 807632
> Change-Id: Id12e3df3073f3b9550d5bc9dfea0068088496f8a
> Reviewed-on: https://pdfium-review.googlesource.com/26910
> Reviewed-by: dsinclair <dsinclair@chromium.org>
> Reviewed-by: Tom Sepez <tsepez@chromium.org>
> Commit-Queue: Nico Weber <thakis@chromium.org>

TBR=thakis@chromium.org,tsepez@chromium.org,dsinclair@chromium.org

Change-Id: I6bcd6c7aadd43aa0a68b9cd7b1d47e10e2c487ae
No-Presubmit: true
No-Tree-Checks: true
No-Try: true
Bug: 807632