Simplify formcalc token list
The keyword list in the formcalc lexer is only used to match
identifiers. We don't need to store the non-identifier tokens in the
list, so they're removed. The hash is removed and the list is compared
by string instead.

The token names have been moved to DEBUG so they won't be included in
Release builds.

Change-Id: Ieec00e9944960e559079083a605e3249c4128841