[formcalc] Do not convert if to a identifier
The if token in formcalc is an internal keyword and can not be used as
an identifier. The current if parsing code will turn if into an
identifier if it fails to find the 'then' statement after the argument
list.

This should be a failed parse instead of coverting if to an identifier
and resetting the lexer.

Change-Id: Ieebf6a1aabc27482fcaeaf7a9bd4be40fc01e9ad