[formcalc] Remove the ParseBlockExpression method
In all the places we call ParseBlockExpression we really mean to parse
an ExpressionList. Using ParseBlockExpression is confusing because the
formcalc grammar has a BlockExpression defined which corresponds to our
ParseDoExpression.

This CL converts the use of ParseBlockExpression into calls to
ParseExpressionList and then builds the BlockExpression class as needed.

Change-Id: Iab9b6e783f01d35df905539d4dbc20ba18a7b0f9