[formcalc] Cleanup if expression parsing.
This CL simplifies the if expression parsing. The elseif construct has
been split into it's own list of IfExpressions and long longer gets
stuffed into the elseExpression. A loop was added into the output, which
also means we can go from having:

if () {
} else {
  if () {
  } else {
  }
}

to
if () {
} else if () {
} else {
}

Which is a bit nicer.

Change-Id: I6edf75215500d6a32a8d7218d477b6960f03de51