[formcalc] Cleanup error handling
This CL cleans up the setting of m_error. In most cases we don't need to
set m_error it will be set when we bubble up the nullptr return from the
various parse methods.

The m_error was set inconsitently previously and was confusing on if it
needed to be set or not.

Change-Id: Id87cc3a909aa94965cb8d2e0a1f58542801660a4