Add convience tags for generating boiler plate in XFA .in files
This adds a number of {{}} style tags to the .in format that allow the
fixup script to inject XFA related boilerplat into the PDF, instead of
having the .in file replicate this. A simple hello wolrd test is also
added to demonstrate this working and as a starting point for
additional XFA tests.

BUG=pdfium:1008

Change-Id: Ie055b6640969ce8291b4c96b401ebf6887dfa0c0