[Merge M65] Fixup AFSimple_Calculate indexing bug
The refactoring in [1] introduced a bug where we pulled the wrong array
index to get the array of fields to use in the AFSimple_Calculate
method. This caused the calculation to not be performed as the fields
could not be found.

This CL fixes up the index and adds a test to verify we use
AFSimple_Calucate to sum two fields.

1- https://pdfium-review.googlesource.com/c/pdfium/+/16670

TBR: rharrision@chromium.org
Bug: chromium:811913
Change-Id: I2abb9fe4fe2d6969ce1876736faadf252ef90575
Reviewed-on: https://pdfium-review.googlesource.com/27670
Commit-Queue: dsinclair <dsinclair@chromium.org>
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
(cherry picked from commit 9141116074f9209f617d9607e747c8947c952c46)