Move XFA pixel tests under pixel/
This removes the need for them to have their own runner. Due to other
upstream changes some of the expectation files needed to be
updated. The suppressions have been updated to make sure the tests do
not run for non-XFA builds.

Having these tests run on the bots surfaced an MSAN issue in the XFA
code, which is fixed in this CL.

BUG=pdfium:1020,pdfium:1008

Change-Id: I44e2fa1cba2d07f30f56f5792749acbe1ebf4b62