[formcalc] Cleanup m_error handling
This CL cleans up the setting of m_error. In most cases we don't need to
set m_error it will be set when we bubble up the nullptr return from the
various parse methods.

The m_error was set inconsitently previously and was confusing on if it
needed to be set or not.

Change-Id: I8648b6296ef15239bd2663e6543a960b88177721