Update patch set 5
Patch Set 5:

> Patch Set 5: Code-Review+1
> 
> (10 comments)
> 
> Got a test case to tickle the new pdfium_test switch?

Any of the javascript/ tests will cause the output. Turning it on the test runner would be possible, but we should hold it for another CL as there will be a bunch of test result churn. I'm sure of a better way to test the pdfium_test change?

Patch-set: 5
Reviewer: Gerrit User 5965 <5965@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Commit-Queue=0