[formcalc] Consider width along with depth of tree
When building the formcalc parser trees we need to limit on width along
with depth. It's possible to generate a tree of a single depth but is
more then 20k nodes wide. This will eventuall cause stack overflow
issues.

This CL re-uses the depth check for the grammar expressions in which
we're extending the width of the tree we count that against our depth
check.

Bug: chromium:813346
Change-Id: I01f6567a75776a75374465eacc1ff546db46cac1