Add test that non-ASCII characters to not foul hyphen processing
In this doc there is a hyphen then a fi ligature. Older versions of
PDFium were not inserting the appropriate 'soft hyphen' marker at this
point, since they were only checking for ASCII characters. That bug is
fixed in the preceding CL, so this test is being added to avoid
regressions.

BUG=pdfium:1029

Change-Id: I9c9906de5a13ade1400a589fb18967938150516d