Update patch set 8
Patch Set 8:

> Patch Set 7:
> 
> Is font_roboto.stream just a .ttf file, or has it been transformed in some way?
> 
> It's a bit big. Assuming it started live as a .ttf file, I wonder if we can subset it with ASCII + some small set of characters, which may be sufficient for testing purposes.

It probably started as a .ttf. I generated the data for it by embedding Roboto-Regular in a simple PDF in Adobe Acrobat and then extracting the resulting font related bits out of it.

I am not honestly not familiar with how the font binary is laid out, so it will probably take me a bit to remove parts of it. Should we have slimming down the font data as a follow on bug/CL?

Patch-set: 8
Reviewer: Gerrit User 7065 <7065@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Commit-Queue=0