Update patch set 7
Patch Set 7:

> Patch Set 5:
> 
> (1 comment)
> 
> > Patch Set 5:
> > 
> > > Patch Set 4:
> > > 
> > > I added a test case using the file from Jane's bug report. With the current API, I believe all I can do is compare the hash of the rendered bitmap, after manually (visually?) checking if that bitmap looks right.
> > > This may change once we have Jane's other CL, 12012: "Updated the annotation APIs to work with multiple sets of quadpoints", which I'll be looking into next.
> > 
> > Ideally, we would have a smaller, possibly handcrafted test case, instead of an 100 KB blob.

The 'blob' renders differently on Mac, resulting in a different md5 hash.
I fixed the issue by defining the hash via
#if _FX_PLATFORM_ == _FX_PLATFORM_APPLE_
...
which is used a lot in other tests, too.

Do you think we can leave this as a test for now, and revisit it later, when the API is in place?

Patch-set: 7
Reviewer: Gerrit User 6485 <6485@fca2d64a-0420-3532-b200-12848c95a47b>