Expand ~ when present in directory path
pdfium_test currently isn't expanding ~ in paths that are passed in
for directories, which is causing commands like this pdfium_test
--font-dir=~/Fonts to fail. This CL adds in performing sh style
expansion on the ~ character for --font-dir and --bin-dir.

This change depends on POSIX wordexp being available, so guarding the
code with checks for Linux || OSX, since those are the only platforms
that we build for that are POSIX complient in this regard.

BUG=pdfium:1027

Change-Id: I05d518dfa71d95c3b0cfd1f4b382e0a6f4a0f515