Annotation API: multiple quadpoints support, replace "attachment points"
The annotation part of the public API only supports adding and retrieval
of a single set of quadpoints, which means the API is limited to
creating one-line text markup annotations (highlight, underline etc.).
This is a serious limitation e.g. for a PDF editor.

This CL, based on Jane Liu's (janeliulwq@google.com) CL
https://pdfium-review.googlesource.com/12012, implements support for
adding and retrieving multiple sets of quadpoints.

The PDF specification and reference documents don't mention the term
"attachment points", so it may cause confusion. All occurrences of that
term were replaced with "QuadPoints".

BUG=pdfium:1045,pdfium:1046

Change-Id: I9e440181df73cdaba01d0f12fe51e31a4e06105f