Add more CPDF_NPageToOneExporter member variables.
Now that CPDF_NPageToOneExporter and CPDF_PageExporter are separate
classes, and it is clear CPDF_NPageToOneExporter is an object designed
to be used once via ExportNPagesToOne(), make some variables declared in
ExportNPagesToOne() class members to reduce the number of parameters in
private methods.

All the member variables are document-level objects. Page-level objects
are still declared in the for-loop inside ExportNPagesToOne().

Change-Id: Ie0b17d5656e5a7232c12aa4061cd58c0844e6f4c