Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I like breaking it down, but moving half of it to another file that's not focused in a functionality or a theme doesn't help much, it just makes searching for a function harder. I'd move the output methods (WritePng, WritePpm, etc.) to a separate file. Annotation-related code could be in another one. Page events in another. Attachments in another.


That's what I did. All the stuff that writes to files is in _write_helper, all the stuff that dumps to console is in dump_helper. The code for event handling is in event_helper.

Patch-set: 1
Reviewer: Gerrit User 5965 <5965@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Commit-Queue=0