Add support for multiple quadpoints
The annotation part of the public API only supports adding and retrieval
of a single set of quadpoints, which means the API is limited to
creating one-line text markup annotations (highlight, underline etc.).
This is a serious limitation e.g. for a PDF editor.

This CL, based on Jane Liu's (janeliulwq@google.com) CL
https://pdfium-review.googlesource.com/12012, implements support for
adding and retrieving multiple sets of quadpoints.

BUG=pdfium:1045

Change-Id: I6265d21cedfd612f0f9857c34665332f71353e3e