Smaller post expression set
This CL decreases the kMaxPostExpressions to 256. This is the number of
accessors you can attach to a single statement (e.g. foo.#A.#A.#A).
Having a very large number can cause stack overflows. The accessor does
not seem like it would expect hundreds of entries on a single element.

Bug: chromium:820688
Change-Id: I19966b43c96f5d1d02a79af127a0c96609420811