Fix XFA caret blinking only while mouse moves.
Currently rect invalidations in XFA are only sent to the embedder
only when RunInvalidate() is executed.
For things which redraw on a timer, such as the caret, there was
no user event to call RunInvalidate() so the page would not redraw.
This CL changes the XFA code to send the invalidations to the
embedder immediately and expects the embedder to combine the
invalidations to limit overdraw.

Bug: chromium:828561
Change-Id: I298052fd7d0c373b029eec191cc6c74c63978348