Make Internal field usage in cfxjs_engine match README.doc
As it turns out, this doesn't cause any bugs with the FXJS/FXJSE
interaction since the magic values will never be present in the
other slot, but the code looks wrong wrt. the document.

Also fix an assert in FXJSE that our objects have two slots,
and null appropriately (just a defensive measure).

Change-Id: I3146fe58350da5e9b76e711d81480565dabd587f