Update patch set 5
Patch Set 5:

Since this is intended to be a 'replacement for pointer usage', what exactly does that mean? What are some examples where ++ / += / + would be useful? At least in Chromium it doesn't seem like we've needed them (and it seems like it would be nice to minimize the divergences, since the span proposal itself is somewhat of a moving target).

Patch-set: 5
Reviewer: Gerrit User 5625 <5625@fca2d64a-0420-3532-b200-12848c95a47b>