Update patch set 2
Patch Set 2:

(5 comments)

I'm OK with this. A few random thoughts, to handle as you'd prefer (I'm OK if the suggestions are declined. As I'm not an OWNER in pdfium, it's not really my place to control the style of pdfium anyway =)

Patch-set: 2
Reviewer: Gerrit User 5625 <5625@fca2d64a-0420-3532-b200-12848c95a47b>