Update patch set 7
Patch Set 7:

I think the right thing to do is to leave span as-is.

What we've got here is a new animal, a "Bounded Pointer", which can be implemented, if desired, as a sub-class of span here in fxcrt.

Patch-set: 7
Reviewer: Gerrit User 5495 <5495@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Commit-Queue=0