Reserve space to reduce memory operations while encoding barcode
In the test case from the bug, the majority of the time is being spent
resizing Widestring internal buffers, since += is being called in a
tight loop. Since the size of the input being mutated and stored is
known, this CL reserves the space before hand to lower thrashing. This
substantially improves runtime of this test case locally.

BUG=chromium:802242

Change-Id: I5176dabc94634b4d6bc3e9425fe6469a5bf35a41