Fix issues with PDFium third_party/base/span.h
Remove stray const in operator[] that was introduced when downgrading
from C++14 to C++11 syntax/

Add missing Get() in first() that was introduced when converting
to UnownedPtr.

Prevent ASAN from flagging spans where the UnownedPtr points to byte
N+1 of a N byte object, and the span is empty. This is legal in C for
ordinary pointers so long as the pointer isn't de-referenced, but is
not allowed per the rules for UnownedPtr. As a side-effect, empty span's
data() method will return nullptr, perhaps catching some other abuses.

Change-Id: Ic143c5ef4e37c1cf86f0a3e5408be6e2076a85e2