Update patch set 11
Patch Set 11:

(2 comments)

I'm not sure if this one is better or not due to the ownership semantics that happen. Could you do something like have the span<> owned by the WideString and it just returns a pointer to the span in GetBuffer?

Patch-set: 11
Reviewer: Gerrit User 5965 <5965@fca2d64a-0420-3532-b200-12848c95a47b>