Update patch set 19
Patch Set 19:

It must be a build system issue unrelated to jumbo. Jumbo is implemented through wrapper files with #include statements. There is no way for it to include the wrong version of the code even if it tried. 

The worst it could theoretically do (which seems to never have happened) is that it could refer to a file that has been removed or not refer to a newly added file.

Patch-set: 19