Update patch set 6
Patch Set 6:

> Did you try clean builds, deleting everything first?

I'm afraid I did, I spent half a day and quite a few CPU cycles building this over and over again, and I still don't understand what's going on. If the line is there, the linker doesn't find the = operator, and if I remove it, it does. Doesn't make sense to me.

What exactly is the purpose of this line?

Patch-set: 6