Remove CPWL_Edit rich text support
This CL removes the code which attempts to provide rich text support for
CPWL_Edit. The code, as written, doesn't actually work well. If you have
tags under the <p> tag we will just lose the content (so
<p><b>foo</b></p> would end up being a blank string). Second, we never
read the "RV" field from the dictionary. The RV field is what actually
stores the rich text value. So, we never set a rich text value into the
text edit.

Change-Id: Idc1d65e1b5d75380dd28c45bcbf9137a2a4cea9a