Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> FPDF_SetSandBoxPolicy() has no existing callers. Can we clarify if it's something that can flip settings on or off at anytime, or if it is more of a initialization phase API?

Good question.  Currently, it can be flipped at any time.  The problem with JS is that once granted, you can't really take it away -- that would mean hunting down timers and killing it. Similarly, turning it on after the fact would require a reload - since there are load actions that are skipped, etc.

And I think we want to be able to do this on a per-document basis even within the same PPAPI process.  So maybe add something to the call when we set up the FFE?

Patch-set: 6
Reviewer: Gerrit User 5495 <5495@fca2d64a-0420-3532-b200-12848c95a47b>