Fix caret moving to next line on click in multiline XFA edit.
On a mouse click targeting right of a line (except for the last one)
in an XFA text edit, the caret was moved to the next line, as the
position after the space was actually in the next line.

This CL changes the behavior so that the caret is positioned before
the space in this case.

Bug: chromium:832293
Change-Id: Ib376e4cca7b32a87d478d3346cb3d8f3bd825daa