Fix caret moving to next line on click in multiline XFA edit.
On a mouse click after the end of a line (except for the last line)
in an XFA text edit, the caret was moved to the beginning of the next
line, as the position after the space was actually in the next line.

This CL changes the behavior so that the caret is positioned before
the space in this case.

Note it's possible to click directly to that position by clicking
on the beginning of the next line.

Bug: chromium:832293
Change-Id: Ib376e4cca7b32a87d478d3346cb3d8f3bd825daa