Remove CFX_BlockBuffer
This CL removes the usage of CFX_BlockBuffer from CFX_XMLParser. The
block buffer has been replaced by a vector which is emptied out after
the characters are removed. This should use less memory when parsing XML
as the block buffer was previously storing all text characters seen in
the file.

Change-Id: I89568c664c762bb9feb034348524e5e86c2d9078