Don't store CE extension data when decoding GIF
GIF extensions are laid out size byte, chunk of data, size byte, chunk
of data, etc, end with a terminator. The decoder needs to scan along
data, finding the size bytes to figure out where the extension ends in
the stream, even if we don't caure about the content. Currently the
decoder is storing all of the data chunks, which are never used and
take up a lot of time to concat together if they are very small.

Our implementation of the GIF does not handle this extension, so when
scanning for the end of the extension, just don't both storing data
from it.

BUG=chromium:833168

Change-Id: Iadf3ab3afd8145b6c5c7c22c30fe9316efcafc15