Don't store CE (comment extension) block data when decoding GIF
GIF extensions are laid out as follows: <size byte> <chunk of data>
<size byte> <chunk of data> ... <terminator byte>. The decoder needs
to scan along the data, finding the size bytes to determine where
the extension block in the stream, even if we don't care about the
content. Currently the decoder is storing all of the data chunks,
which are never used and take up a lot of time to concat together if
they are very small.

Our implementation of the GIF spec does not handle this extension, so
when scanning for the end of the block, just don't bother storing
data from it.

BUG=chromium:833168

Change-Id: Iadf3ab3afd8145b6c5c7c22c30fe9316efcafc15