Avoid perpetual spins for parserless CPDF_Pages
This is an obstacle to introducing dummy CPDF_Pages to underly all
CPDFXFA_Pages, since we would then have a parserless CPDF_Page, whose
parse could never succeed or fail.

Given that we clear the parser on a successful parse, force the caller
to call us again and use that as the sole signal of a successful parse,
which also handles the "never had a parser" case in the same way.

Small separate CL since it might impact non-XFA cases unexpectedly.

Change-Id: I002064e5d4caa44935f50bde0f3b00e3b8cb0bfe