Make default realloc for libs return NULL instead OOM
The current default implementation of realloc being supplied to C libs
will OOM when an oversized allocation occurs. This is unlike calloc
and malloc, which return NULL in this case. Changing realloc to return
NULL.

This also adds in PartitionReallocGenericFlags and has the existing
PartitionReallocGeneric call it with an empty flags value. This is the
same pattern that is used for PartitionAllocGeneric, and allows for
passing in the flag that controls the 'return NULL' vs OOM behaviour.

BUG=chromium:783022

Change-Id: Ib11a79fbdd65c9365d53469d5594addbb3711fa9