[xfa] Verify field count before accessing
When processing items for layout it's possible for the iBlockIndex*3
value could be larger then the field split count. If this is the case
we'll walk off the end of the split array.

This CL verifys that we have enough data before attempting to walk the
splits and returns early if we don't have enough data.

Bug: chromium:837585
Change-Id: I534298b4ee354ce079442d893202f811431155a0