Use checked large integer in ContinueQuickStretch
This existing code has the potential for an integer overflow in it.
When overflow occurs in this function scaling may partially succeed.
This is due to how out of range values are being clamped, which
implicitly swallows the overflow.

This CL changes the calculation to be performed in a 64-bit space and
then attempts to down cast it back to 32-bit space at the end. Because
there are multiple steps it is possible for an intermediate value to
cause an overflow in 32 bit space, but the final value to be valid. If
the downcast is not possible then the stretch operation is failed.

An existing test case has been updated, since it encoded an incorrect
result.

BUG=chromium:839245

Change-Id: I637cc1e2d6c6c2d5394599104f76352c20ead021