Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)
> 
> The way we win at this game is to replace FixedBufGrow with buf[fixed], by determining a bound on fixed and ditching the grow.
> 
> Can we set a max number of components somewhere and just flunk things that exceed it?

I think that for most cases the buffer is of a small size. However there are some cases where the number can be large while still being a legitimate PDF. For example, looking at CPDF_DeviceNCS::v_Load (which return the number assigned to m_nComponents), it returns the number of objects in an array. Per PDF spec 'DeviceN colorspaces can contain an arbitrary number of color components'. So the problem of buf[fixed] is that it would need to fit the stack and there is no reasonable bound on fixed.

Patch-set: 1
Reviewer: Gerrit User 6295 <6295@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Commit-Queue=0