Update patch set 4
Patch Set 4:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > > Patch Set 3: Code-Review-1
> > > 
> > > This restores the file to an unreadable mess. I'd rather figure out why the date widget isn't drawing then to restore the bad code.
> > 
> > It is, but the Rect that's its parent thinks it's a TextEdit.
> 
> 
> Ok, so why does the parent think it's a TextEdit instead of the correct type? That seems like the bug we need to track down, yea?

In this case, we have a Draw which has a Rectangle as first child and a DefaultUI as uiChild. The current code asks the uiChild (which is a DefaultUI) for what the type of widget it is, and GetDefaultFFWidgetType() always returns TextEdit. This value is not overridden because it's set already, and is eventually returned, which we should be getting Rectangle back.

I think the current logic doesn't work, at least in this case, but I couldn't find a simpler way to fix than restoring the previous logic.

Added a test as we talked offline. I'll rename the variables to the new naming.

Patch-set: 4
Reviewer: Gerrit User 7035 <7035@fca2d64a-0420-3532-b200-12848c95a47b>